requirements.txt
fastapi
uvicorn
streamlit
faiss-cpu
sentence-transformers
python-dotenv
pillow
pytesseract
SpeechRecognition
google-genai
